DFK-598 - Fix etter CR, flyttet logikk for å hente styringsdata til testing-api.ts
